src/requirements.txt
mtcnn
matplotlib
sklearn
pillow
split_folders
